Error corection

+++ lesson_1/ideal_weight.rb
@@ -3,7 +3,7 @@ name = gets.chomp
 name.capitalize!
 
 print "Какой у вас рост? "
-weight = gets.chomp.to_i
+weight = gets.chomp.to_f
 
 ideal_weight = weight - 110
 if ideal_weight  >= 0
